PoC loader
We want to be able to load / require('./mytemplate.njk') and have it
bundle all dependent templates.

We want it to play nice with other loaders (specifically html-loader) so
that we can rely on referencing dependencies in templates via html tags.

We want it to parse the templates during webpacking so that we can catch
errors early.

How this works

The "pitch" stage is used to parse the templates and build the list of
dependent templates and output a file with a set of requires for each
template source.

We let the require mechanism be the source of truth for the templates
source so that we do not duplicate templates accross multiple requires.