add a way of testing various webpack configs
We want to be able to test various configuration scenarios to ensure
that the loader works in different setups and different environments.

A _very_ basic test framework was created to allow a set of example
configurations to be tested, and these are build against both
target:node (and tested in ndoe) and target:web (tested via browser-run)

Notable changes:

* Templates are precompiled where possible by default.
* When the loader detects that it is not the first loader in the chain it
will fallback to non-precompiled templates.
* When non-compiled templates are in use the non-SLIM version of the
runtime will be bundled.
* Bundling for target:web now works as expected.