precompile options
precompile: 'auto' - this is the default value, it will try it's best to
precopmile templates and bundle the 'slim' version of the nunjucks
runtime. Most of the time this is what you want.

precompile: true - if this is enabled it will throw an error during
webpack-time if it is unable to precompile templates and use slim. This
option is useful to protect against accidently introducing webpack
configuration that triggers the non-precompiled bundling which may be
important for browser bundles

precompile: false - if this option is set, precompilation is never
attempted and the full runtime is bundled. The only reason I can think
of that you might want this option is if you want to post process
nunjucks with some kind of loader that expects the raw source.