Fix issue imports based on a node_modules path
If you are attempting to import a njk template from a node_module ie

```
{% from "my-module/x.njk" import y %}
```

then the parentPath given to the template Loader would be a relative
path and not the expected absolute path required to reference the
template.

To work around this, we pass the expected name when compiling the
compiled ahead of time so the paths match.